Add low-level routine for 2-d definite Gaussian integrals
This is a translation from a matlab routine, for which there is also
a FORTRAN routine.  We include the FORTRAN version here, but
only as a reference implementation used to generate test data
for comparison, so we don't have to rely on being able to link
FORTRAN and C++ on arbitrary platforms.