Don't use afw makeLocalWcs - we'll just implement it here.
makeLocalWcs was added to afw on a branch I was building against, but
it turned out to be so trivial (and likely not as globally useful as I originally
thought it might be).  So we'll just implement it here and abandon the code
in afw instead of merging it to afw master.