Fix bugs in SoftenedLinearPrior
A review comment that a factor of 2 should be a factor of PI uncovered a host
of other problems, which are now more completely tested.  Aside from the
integral normalization being wrong (which wouldn't have mattered for any
current use of this code), we also had a DC offset in the middle of the prior,
which made the softening not really soften the edges.