Remove persistence support from custom record/table classes.
The custom records and tables here are deprecated and would be
removed entirely if our unit tests for other components didn't rely
on them.  Those unit tests don't require persistence, and hence it's
simplest to just remove persistence instead of updating it in response
to changes in the persistence base classes in afw.