Switch to new fit region determination in CModel.
After this change, CModel will use the shape slot
and Kron radius (if available) to set the initial fit
region instead of the detection Footprint, and it
will put an upper bound as well as a lower bound
on the final fit region from the initial fit.  The
net result seems to be greatly improved computational
performance and a small decrease in the number of
spuriously large objects.