Add error flag handling to PsfFitter, using a wrapper PsfFitterAlgorithm.
This class adds measure and fail to the original fitter, and also monitors
the optimizer state and checks for failures.

This is not a child of SimpleAlgorithm, which has a different control type.
Also, the python code which exposes this algorithm can call multiple fitters
in sequence, so it is really multiple algorithms. But each fitter has separate
measure and fail methods, and also its own set of flags. There is nt global flag.
The fitters themself fitters call fail, not the python task.

Also, add unit test which tests the error flag setting. (Except for Exception)