UnitTransformedLikelihood: always include some kind of weight.
In Bayesian fitting, we need to do some kind of variance weighting in
the likelihood even when we want a constant weight across all pixels -
otherwise the likelihood and prior will not be weighted correctly
relative to each other.

Note that this makes it impossible to run CModel on an image with
exactly zero variance.